+++ src/test/scala/org/broadinstitute/dsde/workbench/sam/api/ResourceRoutesV1Spec.scala
@@ -9,6 +9,7 @@ import org.broadinstitute.dsde.workbench.model.ErrorReportJsonSupport._
 import org.broadinstitute.dsde.workbench.model.WorkbenchIdentityJsonSupport._
 import org.broadinstitute.dsde.workbench.model._
 import org.broadinstitute.dsde.workbench.sam.TestSupport.{genGoogleSubjectId, _}
+import org.broadinstitute.dsde.workbench.sam.api.TestSamRoutes.SamResourceActionPatterns
 import org.broadinstitute.dsde.workbench.sam.model.SamJsonSupport._
 import org.broadinstitute.dsde.workbench.sam.model._
 import org.broadinstitute.dsde.workbench.sam.openam.MockAccessPolicyDAO
@@ -29,40 +30,6 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
   val defaultUserInfo = UserInfo(OAuth2BearerToken("accessToken"), WorkbenchUserId("user1"), WorkbenchEmail("user1@example.com"), 0)
 
   private val managedGroupResourceType = configResourceTypes.getOrElse(ResourceTypeName("managed-group"), throw new Error("Failed to load managed-group resource type from reference.conf"))
-
-  private object SamResourceActionPatterns {
-    val readPolicies = ResourceActionPattern("read_policies", "", false)
-    val alterPolicies = ResourceActionPattern("alter_policies", "", false)
-    val delete = ResourceActionPattern("delete", "", false)
-
-    val sharePolicy = ResourceActionPattern("share_policy::.+", "", false)
-    val readPolicy = ResourceActionPattern("read_policy::.+", "", false)
-
-    val setPublic = ResourceActionPattern("set_public", "", false)
-    val setPolicyPublic = ResourceActionPattern("set_public::.+", "", false)
-
-    val use = ResourceActionPattern("use", "", true)
-    val readAuthDomain = ResourceActionPattern("read_auth_domain", "", true)
-
-    val testActionAccess = ResourceActionPattern("test_action_access::.+", "", false)
-  }
-
-  private val resourceTypeAdmin = ResourceType(
-    ResourceTypeName("resource_type_admin"),
-    Set(
-      SamResourceActionPatterns.alterPolicies,
-      SamResourceActionPatterns.readPolicies,
-      SamResourceActionPatterns.sharePolicy,
-      SamResourceActionPatterns.readPolicy,
-      SamResourceActionPatterns.setPublic,
-      SamResourceActionPatterns.setPolicyPublic
-    ),
-    Set(
-      ResourceRole(
-        ResourceRoleName("owner"),
-        Set(SamResourceActions.alterPolicies, SamResourceActions.readPolicies, SamResourceActions.setPublic))),
-    ResourceRoleName("owner")
-  )
 
   private val defaultTestUser =  CreateWorkbenchUser(WorkbenchUserId("testuser"), genGoogleSubjectId(), WorkbenchEmail("testuser@foo.com"), None)
 
@@ -795,10 +762,9 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.alterPolicies, SamResourceActions.readPolicies))),
       ResourceRoleName("owner"))
 
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
-    runAndWait(samRoutes.resourceService.addSubjectToPolicy(model.FullyQualifiedPolicyId(model.FullyQualifiedResourceId(resourceTypeAdmin.name, ResourceId(resourceType.name.value)), AccessPolicyName("owner")), samRoutes.userInfo.userId, samRequestContext))
+    runAndWait(samRoutes.resourceService.addSubjectToPolicy(model.FullyQualifiedPolicyId(model.FullyQualifiedResourceId(TestSamRoutes.resourceTypeAdmin.name, ResourceId(resourceType.name.value)), AccessPolicyName("owner")), samRoutes.userInfo.userId, samRequestContext))
 
     val resourceId = ResourceId("foo")
     val policyName = AccessPolicyName("bar")
@@ -976,9 +942,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readPolicies))),
       ResourceRoleName("owner"))
 
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -996,9 +960,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readPolicy(AccessPolicyName("owner"))))),
       ResourceRoleName("owner"))
 
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -1016,9 +978,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.delete))),
       ResourceRoleName("owner"))
 
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -1035,11 +995,10 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.alterPolicies, SamResourceActions.readPolicies))),
       ResourceRoleName("owner"))
 
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
     runAndWait(samRoutes.resourceService.addSubjectToPolicy(model.FullyQualifiedPolicyId(
-        model.FullyQualifiedResourceId(resourceTypeAdmin.name, ResourceId(resourceType.name.value)), AccessPolicyName(resourceTypeAdmin.ownerRoleName.value)), samRoutes.userInfo.userId, samRequestContext))
+        model.FullyQualifiedResourceId(TestSamRoutes.resourceTypeAdmin.name, ResourceId(resourceType.name.value)), AccessPolicyName(TestSamRoutes.resourceTypeAdmin.ownerRoleName.value)), samRoutes.userInfo.userId, samRequestContext))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -1073,9 +1032,8 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       ResourceRoleName("owner")
     )
 
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
     runAndWait(samRoutes.resourceService.addSubjectToPolicy(model.FullyQualifiedPolicyId(
         model.FullyQualifiedResourceId(resourceTypeAdmin.name, ResourceId(resourceType.name.value)), AccessPolicyName(resourceTypeAdmin.ownerRoleName.value)), samRoutes.userInfo.userId, samRequestContext))
 
@@ -1094,11 +1052,10 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readPolicies))),
       ResourceRoleName("owner"))
 
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
     runAndWait(samRoutes.resourceService.addSubjectToPolicy(model.FullyQualifiedPolicyId(
-        model.FullyQualifiedResourceId(resourceTypeAdmin.name, ResourceId(resourceType.name.value)), AccessPolicyName(resourceTypeAdmin.ownerRoleName.value)), samRoutes.userInfo.userId, samRequestContext))
+        model.FullyQualifiedResourceId(TestSamRoutes.resourceTypeAdmin.name, ResourceId(resourceType.name.value)), AccessPolicyName(TestSamRoutes.resourceTypeAdmin.ownerRoleName.value)), samRoutes.userInfo.userId, samRequestContext))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -1115,9 +1072,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.alterPolicies, SamResourceActions.readPolicies))),
       ResourceRoleName("owner"))
 
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -1137,9 +1092,8 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readAuthDomain, ManagedGroupService.useAction))),
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin, managedGroupResourceType.name -> managedGroupResourceType))
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, managedGroupResourceType.name -> managedGroupResourceType))
 
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
     runAndWait(samRoutes.managedGroupService.createManagedGroup(ResourceId(authDomain), defaultUserInfo, samRequestContext = samRequestContext))
 
     val resourceId = ResourceId("foo")
@@ -1161,9 +1115,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readAuthDomain, ManagedGroupService.useAction))),
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin, managedGroupResourceType.name -> managedGroupResourceType))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, managedGroupResourceType.name -> managedGroupResourceType))
 
     val resourceId = ResourceId("foo")
     val policiesMap = Map(AccessPolicyName("ap") -> AccessPolicyMembership(Set(defaultUserInfo.userEmail), Set(SamResourceActions.readAuthDomain, ManagedGroupService.useAction), Set(ResourceRoleName("owner"))))
@@ -1185,9 +1137,8 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(ManagedGroupService.useAction))),
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin, managedGroupResourceType.name -> managedGroupResourceType))
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, managedGroupResourceType.name -> managedGroupResourceType))
 
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
     runAndWait(samRoutes.managedGroupService.createManagedGroup(ResourceId(authDomain), defaultUserInfo, samRequestContext = samRequestContext))
 
     val resourceId = ResourceId("foo")
@@ -1209,9 +1160,8 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readAuthDomain, ManagedGroupService.useAction))),
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin, managedGroupResourceType.name -> managedGroupResourceType))
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, managedGroupResourceType.name -> managedGroupResourceType))
 
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
     runAndWait(samRoutes.managedGroupService.createManagedGroup(ResourceId(authDomain), defaultUserInfo, samRequestContext = samRequestContext))
 
     val resourceId = ResourceId("foo")
@@ -1237,9 +1187,8 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readAuthDomain, ManagedGroupService.useAction))),
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin, managedGroupResourceType.name -> managedGroupResourceType))
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, managedGroupResourceType.name -> managedGroupResourceType))
 
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
     runAndWait(samRoutes.managedGroupService.createManagedGroup(ResourceId(authDomain), defaultUserInfo, samRequestContext = samRequestContext))
 
     val resourceId = ResourceId("foo")
@@ -1278,9 +1227,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readPolicies))),
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -1301,9 +1248,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readAuthDomain))), // any action except read_policies
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -1320,9 +1265,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readPolicies))),
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
@@ -1343,9 +1286,7 @@ class ResourceRoutesV1Spec extends FlatSpec with Matchers with ScalatestRouteTes
       Set(ResourceRole(ResourceRoleName("owner"), Set(SamResourceActions.readPolicies))),
       ResourceRoleName("owner")
     )
-    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType, resourceTypeAdmin.name -> resourceTypeAdmin))
-
-    samRoutes.resourceService.initResourceTypes().unsafeRunSync()
+    val samRoutes = TestSamRoutes(Map(resourceType.name -> resourceType))
 
     val resourceId = ResourceId("foo")
     runAndWait(samRoutes.resourceService.createResource(resourceType, resourceId, samRoutes.userInfo, samRequestContext))
